requirements.txt
numpy>=1.24.0
networkx>=3.0
openai>=1.0.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
pandas>=2.0.0
pydantic>=2.0.0
pathlib
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
ipykernel>=6.25.0
python-dotenv>=1.0.0
scipy>=1.10.0
tabulate

